# HTML5 parser optimizations