--- a/ext/dom/tests/modern/html/parser/user_error_handler.phpt
+++ b/ext/dom/tests/modern/html/parser/user_error_handler.phpt
@@ -1,0 +1,19 @@
+--TEST--
+Parse HTML document with user error handler and error_reporting(0)
+--EXTENSIONS--
+dom
+--INI--
+error_reporting=0
+--FILE--
+<?php
+
+set_error_handler(function ($errno, $errstr, $errfile, $errline) {
+    var_dump($errno, $errstr);
+}, E_WARNING);
+
+DOM\HTMLDocument::createFromString('<html></html>');
+
+?>
+--EXPECT--
+int(2)
+string(113) "DOM\HTMLDocument::createFromString(): tree error unexpected-token-in-initial-mode in Entity, line: 1, column: 2-5"