--- a/ext/libxml/php_libxml.h
+++ b/ext/libxml/php_libxml.h
@@ -163,6 +163,7 @@ PHP_LIBXML_API void php_libxml_issue_error(int level, const char *msg);
 PHP_LIBXML_API bool php_libxml_disable_entity_loader(bool disable);
 PHP_LIBXML_API void php_libxml_set_old_ns(xmlDocPtr doc, xmlNsPtr ns);
 PHP_LIBXML_API php_stream_context *php_libxml_get_stream_context(void);
+PHP_LIBXML_API bool php_libxml_uses_internal_errors(void);
 
 PHP_LIBXML_API zend_string *php_libxml_sniff_charset_from_string(const char *start, const char *end);
 PHP_LIBXML_API zend_string *php_libxml_sniff_charset_from_stream(const php_stream *s);